+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 2026-May-28 (2026-May-29 in UTC)
+
+### Version 0.0.4
+
+- Updated `dis` crate to v. `0.0.2` (renamed `.spanned(span)` --> `.span_err(span)`).
+- Made a few types derive `Clone`.
+
 ## 2026-May-27
 
 ### Version 0.0.3
@@ -11,7 +18,7 @@ Renamed `pass_through_tags` -> `pass_through_mce_tags`.
 ### Version 0.0.2
 
 - Updated docs.
-- Documentation on docs.rs now shows full functionality (that is, including `std` feature.)
+- Documentation on docs.rs now shows full functionality (that is, including `std` feature).
 
 ### Version 0.0.1